lab5/requirements.txt
matplotlib
torch
scikit_image
numpy
torchvision
pillow
